working with db moved to another file
