+++ bot.py
@@ -1,39 +1,10 @@
-
+from repositories import get_user_city, save_user_city
 from settings import BOT_TOKEN
 from logic import get_weather
 
 from telebot import TeleBot
 
 bot = TeleBot(BOT_TOKEN)
-
-# TODO: migrate SQLite
-default_city = 'Yerevan'
-cities = {}
-
-import sqlite3
-
-
-def get_user_city(chat_id):
-    connection = sqlite3.connect("database.db")
-    cursor = connection.cursor()
-    cursor.execute("SELECT city FROM user_cities WHERE chat_id = (?)", (chat_id,))
-    city = cursor.fetchone()
-    return city[0] if city else default_city
-
-
-def save_user_city(chat_id, city):
-    connection = sqlite3.connect("database.db")
-    cursor = connection.cursor()
-
-    cursor.execute("SELECT city FROM user_cities WHERE chat_id = (?)", (chat_id,))
-    result = cursor.fetchone()
-
-    if not result:
-        cursor.execute("INSERT INTO user_cities VALUES (?, ?)", (chat_id, city))
-    else:
-        cursor.execute("UPDATE user_cities SET city = (?) WHERE chat_id = (?)", (city, chat_id))
-
-    connection.commit()
 
 
 @bot.message_handler(commands=['get_weather'])
